Return Option<String> from GetToken::get_token instead of Result<...>
