+++ src/generator/templates/api/lib/mbuild.mako
@@ -711,24 +711,13 @@ else {
         loop {
             % if default_scope:
             let token = match ${auth_call}.get_token(&self.${api.properties.scopes}.iter().map(String::as_str).collect::<Vec<_>>()[..]).await {
-                // TODO: remove Ok / Err branches
-                Ok(Some(token)) => token.clone(),
-                Ok(None) => {
-                    let err = oauth2::Error::OtherError(anyhow::Error::msg("unknown error occurred while generating oauth2 token"));
-                    match dlg.token(&err) {
+                Some(token) => token.clone(),
+                None => {
+                    match dlg.token() {
                         Some(token) => token,
                         None => {
                             ${delegate_finish}(false);
-                            return Err(client::Error::MissingToken(err))
-                        }
-                    }
-                }
-                Err(err) => {
-                    match dlg.token(&err) {
-                        Some(token) => token,
-                        None => {
-                            ${delegate_finish}(false);
-                            return Err(client::Error::MissingToken(err))
+                            return Err(client::Error::MissingToken);
                         }
                     }
                 }
